MMT-2313: moved loss_report_samples_helper.rb and added it to the rails_helper.rb
